Rename CommandLine to Service
